refactor + build out testing flow
