+++ dbt/task/run.py
@@ -1,3 +1,6 @@
+
+import os
+from dbt.templates import BaseCreateTemplate
 from dbt.runner import Runner
 
 class RunTask:
@@ -5,6 +8,11 @@ class RunTask:
         self.args = args
         self.project = project
 
+    def get_target(self):
+        return os.path.join(self.project['target-path'], BaseCreateTemplate.dir())
+
     def run(self):
-        runner = Runner(self.project, self.project['run-target'])
+        target_path = self.get_target()
+
+        runner = Runner(self.project, target_path())
         runner.run()
